+++ examples/servers-within-master-slave-server-group/main.tf
@@ -37,10 +37,19 @@ resource "alicloud_vswitch" "default" {
 
 
 // ECS Module
-module "ecs_instance" {
+module "ecs-instance-master" {
   source                      = "alibaba/ecs-instance/alicloud//modules/x86-architecture-general-purpose"
   region                      = var.region
-  number_of_instances         = 2
+  instance_type_family        = "ecs.g6"
+  vswitch_id                  = length(data.alicloud_vswitches.default.ids) > 0 ? data.alicloud_vswitches.default.ids.0 : concat(alicloud_vswitch.default.*.id, [""])[0]
+  security_group_ids          = data.alicloud_security_groups.default.ids
+  associate_public_ip_address = true
+  internet_max_bandwidth_out  = 10
+}
+
+module "ecs-instance-slave" {
+  source                      = "alibaba/ecs-instance/alicloud//modules/x86-architecture-entry-level"
+  region                      = var.region
   instance_type_family        = "ecs.g6"
   vswitch_id                  = length(data.alicloud_vswitches.default.ids) > 0 ? data.alicloud_vswitches.default.ids.0 : concat(alicloud_vswitch.default.*.id, [""])[0]
   security_group_ids          = data.alicloud_security_groups.default.ids
@@ -54,17 +63,16 @@ module "slb" {
   region = var.region
   servers_of_master_slave_server_group = [
     {
-      server_ids  = concat(module.ecs_instance.this_instance_id, [""])[0]
+      server_ids  = join(",", module.ecs-instance-master.this_instance_id)
       port        = "80"
       weight      = "100"
       type        = "ecs"
       server_type = "Master"
     },
     {
-      server_ids  = concat(module.ecs_instance.this_instance_id, ["", ""])[1]
+      server_ids  = join(",", module.ecs-instance-slave.this_instance_id)
       port        = "80"
-      weight      = "100"
-      type        = "ecs"
+      weight      = "90"
       server_type = "Slave"
     },
   ]